refactor(config): split config.ts into lib/config/{loader,includes,validate,errors}.ts
Mechanical split along existing landmarks: config.ts now re-exports the
public surface (loadWebreelConfig, validateWebreelConfig, etc.) from
focused modules. No behavior changes; config.test.ts assertions unchanged.
